move socket gen logic to SocketContext file
