--- src/containers/shared/SocketContext.js
@@ -1,5 +1,0 @@
-import React from 'react'
-
-const SocketContext = React.createContext(undefined)
-
-export default SocketContext
